Move BloomTradeClient::ExchangeRates::Convert to BloomTradeClient::Convert
