--- app/services/bloom_trade_client/exchange_rates/convert.rb
@@ -1,67 +1,0 @@
-module BloomTradeClient
-  module ExchangeRates
-    class Convert
-
-      def self.call(
-        base_currency:,
-        counter_currency:,
-        reserve_currency: BloomTradeClient.configuration.reserve_currency,
-        type: "mid",
-        jwt:
-      )
-        rate = direct_rate(type, base_currency, counter_currency, jwt)
-        return rate if rate
-
-        origin_rate = direct_rate(type, base_currency, reserve_currency, jwt)
-        destination_rate = direct_rate(type, reserve_currency, counter_currency, jwt)
-        reverse_rate = origin_rate.rate * destination_rate.rate if origin_rate && destination_rate
-        return ConvertResult.new(rate: reverse_rate) if reverse_rate
-
-        Rails.logger.error(
-          "Unable to calculate rate #{base_currency}#{counter_currency}"
-        )
-        return ConvertResult.new(rate: 0.0)
-      end
-
-      private
-
-      def self.direct_rate(type, base_currency, counter_currency, jwt)
-        return nil unless %w(buy sell mid).include? type
-        return ConvertResult.new(rate: 1.0) if base_currency == counter_currency
-
-        resolution = ExchangeRate.resolve(
-          base_currency: base_currency,
-          counter_currency: counter_currency,
-          jwt: jwt,
-        )
-
-        return unless resolution[:rate]
-
-        rate = resolution[:rate]
-
-        if rate.expired?
-          raise(
-            BloomTradeClient::ExpiredRateError,
-            BloomTradeClient::ExpiredRateError.message_from(
-              rate.base_currency,
-              rate.counter_currency,
-              rate.expires_at,
-            )
-          )
-        end
-
-        if resolution[:direction] == "direct"
-          return ConvertResult.new(
-            rate: rate.send(type.to_sym),
-            expires_at: rate.expires_at,
-          )
-        else
-          return ConvertResult.new(
-            rate: 1.0 / rate.send(type.to_sym),
-            expires_at: rate.expires_at,
-          )
-        end
-      end
-    end
-  end
-end
